+++ R/wlsPower.R
@@ -48,7 +48,7 @@
 #' If only one element is provided, autocorrelation of all random effects is
 #' assumed to be the same.
 #' *Currently not compatible with `rho`!=0 !*
-#' @param rho numeric (scalar), correlation of `tau` and `eta`
+#' @param rho numeric (scalar), correlation of `tau` and `eta`. The default is no correlation.
 #' @param gamma numeric (scalar), random time effect
 #' @param psi numeric (scalar), random subject specific intercept.
 #' Leads to a closed cohort setting
@@ -71,6 +71,7 @@
 #' @param dfAdjust character, one of the following: "none","between-within",
 #' "containment", "residual".
 #' @param verbose integer, how much information should the function return?
+#' See also under `Value`.
 #' @param period numeric (scalar)
 #' @param CovMat numeric, a positive-semidefinite matrix with
 #' (#Clusters \eqn{\cdot} timepoints) rows and columns. If `CovMat` is given,
@@ -79,6 +80,9 @@
 #' @param INDIV_LVL logical, should the computation be conducted on an
 #' individual level? This leads to longer run time and is
 #' mainly for diagnostic purposes.
+#' @param INFO_CONTENT logical, should the information content of cluster cells be
+#' computed? The default is `TRUE` for designs with less or equal than 2500
+#' cluster cells, otherwise `FALSE`. Ignored if `verbose=0`.
 #'
 #' @details
 #' Let \eqn{\theta:= \mu_1-\mu_0} the treatment effect under investigation.
@@ -92,9 +96,16 @@
 #' @return
 #' The return depends on the `verbose` parameter.
 #' If `verbose`=0, only the power is returned
-#' If `verbose`=1 (the default), a list containing power and the
+#' If `verbose`=1 (the default), a list containing power, projection matrix and the
 #' parameters of the specific setting is returned.
-#' If requested (by `verbose`=2) this list also contains relevant matrices.
+#' If explicitly requested (by `verbose`=2) this list also contains
+#' the `DesMat`-object and the covariance matrix.
+#'
+#' If INFO_CONTENT= TRUE, the returned list contains a named list with four elements:
+#' `Cells` is explicit computation of the information content in each cell;
+#' `Cluster` is the information content of entire clusters;
+#' `time` is thie information content of entire time periods and
+#' `Closed` is a formula-based computation the information content in each cell,
 #'
 #' @export
 #'
@@ -199,10 +210,12 @@
 #' wlsPower(mu0=0.5, mu1=0.25, Cl=rep(4,8), tau=0.5, N=1,
 #'               family="binomial", marginal_mu=TRUE)
 #'
+#'##
+#'##
+#'
 
 
-
-wlsPower <- function( Cl            = NULL,
+ wlsPower <- function( Cl            = NULL,
                       timepoints    = NULL,
                       DesMat        = NULL,
                       trtDelay      = NULL,
@@ -229,6 +242,7 @@ wlsPower <- function( Cl            = NULL,
                       sig.level     = 0.05,
                       dfAdjust      = "none",
                       INDIV_LVL     = FALSE,
+                      INFO_CONTENT  = NULL,
                       verbose       = 1){
   ## Match string inputs ####
   ### dsntype
@@ -290,12 +304,6 @@ wlsPower <- function( Cl            = NULL,
       if( (-1)>rho | rho>1 )
         stop("Correlation rho must be between -1 and 1")
     }
-    if(is.null(tau)){
-      tau <- 0
-      warning("Random cluster effect tau and random treatment effect eta",
-              " are assumed to be 0, i.e. the observations across clusters are",
-              " assumed to be marginally independent. Declare tau=0 to supress this warning.")
-    }
     if(!is.null(psi) & is.null(power)){
       if(is.null(N))
         stop("If the standard deviation `psi` is not null, N is needed.")
@@ -333,19 +341,20 @@ wlsPower <- function( Cl            = NULL,
                                   timepoints = timepoints,
                                   timeAdjust = timeAdjust,
                                   period     = period,
+                                  incomplete = incomplete,
                                   N          = if(INDIV_LVL) N,
                                   INDIV_LVL  = INDIV_LVL )
   }else{
     if(inherits(DesMat, "DesMat")) {
-      if(!all(sapply(list(Cl, timepoints, trtDelay,
-                          period),     is.null)))      ## timeAdjust??
+      if(!all(sapply(list(Cl, timepoints ,trtDelay,period,incomplete),is.null)))      ## timeAdjust defaults to factor (!)
         warning("If input to argument DesMat inherits class `DesMat`, \n",
-                "Cl, timepoints, trtDelay, ",
+                "Cl, timepoints, trtDelay, incomplete,",
                 "timeAdjust, period and dsntype are ignored.")
     } else if(inherits(DesMat,"matrix") & !inherits(DesMat,"DesMat")){
       DesMat <- construct_DesMat(trtmatrix  = DesMat,
                                  timeAdjust = timeAdjust,
                                  period     = period,
+                                 incomplete = incomplete,
                                  N          = if(INDIV_LVL) N,
                                  INDIV_LVL  = INDIV_LVL)
       if(!all(sapply(list(Cl, timepoints, trtDelay, dsntype), is.null)))
@@ -360,8 +369,17 @@ wlsPower <- function( Cl            = NULL,
   ## declare temporary variables #####
   timepoints <- DesMat$timepoints
   lenCl      <- length(DesMat$Cl)
-  SumCl      <- sum(DesMat$Cl)
+  sumCl      <- sum(DesMat$Cl)
 
+  ## default for INFO_CONTENT ####
+  if(is.null(INFO_CONTENT)){
+    INFO_CONTENT <- ifelse(length(DesMat$trtMat)<=2500 & sumCl>2 &
+                           verbose>0 & !INDIV_LVL,                   TRUE,FALSE)
+  }else if(INFO_CONTENT & sumCl<=2) {
+    INFO_CONTENT <- FALSE
+    warning("Information Content not (yet) implemented for only two clusters.",
+            "INFO_CONTENT set to FALSE.")
+  }
 
   ## distribution family ####
   if(family =="gaussian"){
@@ -381,7 +399,7 @@ wlsPower <- function( Cl            = NULL,
       print(paste("mu0=",round(mu0,5),", mu1=",round(mu1,5),"."))
     }
 
-    muMat   <- matrix(mu0, SumCl, timepoints) + DesMat$trtMat*(mu1-mu0)
+    muMat   <- matrix(mu0, sumCl, timepoints) + DesMat$trtMat*(mu1-mu0)
     sigma   <- sqrt(muMat * (1-muMat))
 
     if (verbose>0) {
@@ -404,40 +422,11 @@ wlsPower <- function( Cl            = NULL,
 
 
   ## incomplete designs #####
-  if(!is.null(incomplete) & is.null(CovMat)){
+  if(!is.null(DesMat$incompMat) & is.null(CovMat)){
+    IM <- DesMat$incompMat
+    IM[IM==0] <- Inf
 
-    if(is.vector(incomplete)){
-      if(dsntype !="SWD")
-        stop("scalar input for argument `incomplete` is only, ",
-             "applicable for dsntype = 'SWD'. ")
-      if(length(incomplete)!=1)
-        stop("incomplete cannot be a vector of length > 1.")
-      if(incomplete>timepoints) {
-        incomplete <- timepoints
-        warning("Argument `incomplete` must be less or equal to the number of",
-                "timepoints. `incomplete` is set to ", timepoints )
-      }
-      Toep <- toeplitz(c(rep(1,incomplete),rep(Inf,lenCl-incomplete)))
-      lastCols <- (timepoints-lenCl+1):timepoints
-
-      IM <- matrix(1,lenCl,timepoints)
-      IM[lower.tri(IM)]                       <- Toep[lower.tri(Toep)]
-      IM[,lastCols][upper.tri(IM[,lastCols])] <- Toep[upper.tri(Toep)]
-
-      IM <- IM[rep(seq_len(lenCl),DesMat$Cl),]
-
-
-    }else if(is.matrix(incomplete)){
-      if(!nrow(incomplete) %in% c(lenCl,SumCl) | ncol(incomplete)!=timepoints)
-        stop("matrix dimensions of argument `incomplete` are ",
-             paste(dim(incomplete),collapse="x"), " but must be ",
-             paste(dim(DesMat$trtMat),collapse="x"), " or ",
-             paste(dim(unique(DesMat$trtMat)),collapse="x"))
-      IM <- incomplete
-      IM[which(IM==0)] <- Inf
-      if(nrow(incomplete)==lenCl) IM <- IM[rep(seq_len(lenCl),DesMat$Cl),]
-    }
-    sigma <- matrix(sigma, nrow=SumCl, ncol=timepoints,
+    sigma <- matrix(sigma, nrow=sumCl, ncol=timepoints,
                     byrow=ifelse(length(sigma)!=timepoints,TRUE,FALSE)) * IM
   }
 
@@ -459,6 +448,7 @@ wlsPower <- function( Cl            = NULL,
                                                            sig.level = sig.level,
                                                            CovMat    = CovMat,
                                                            INDIV_LVL = INDIV_LVL,
+                                                           INFO_CONTENT = FALSE,
                                                            verbose   = 0)},
                 interval=N_range)$root),
               error=function(cond){
@@ -483,6 +473,7 @@ wlsPower <- function( Cl            = NULL,
                           sig.level = sig.level,
                           CovMat    = CovMat,
                           INDIV_LVL = INDIV_LVL,
+                          INFO_CONTENT = INFO_CONTENT,
                           verbose   = verbose)
   if(!is.null(power)) out$N_opt <- N_opt
 
@@ -536,17 +527,18 @@ compute_wlsPower <- function(DesMat,
                              dfAdjust   = "none",
                              sig.level  = .05,
                              INDIV_LVL  = FALSE,
+                             INFO_CONTENT = FALSE,
                              verbose    = 1){
-  dsnmatrix  <- DesMat$dsnmatrix
-  timepoints <- DesMat$timepoints
-  SumCl      <- sum(DesMat$Cl)
+  dsn        <- DesMat$dsnmatrix
+  tp         <- DesMat$timepoints
+  sumCl      <- sum(DesMat$Cl)
   SumSubCl   <- sum(DesMat$N)
   trtMat     <- DesMat$trtMat
 
   ## get covariance matrix #####
   if(is.null(CovMat))
-    CovMat   <- construct_CovMat(SumCl      = SumCl,
-                                 timepoints = timepoints,
+    CovMat   <- construct_CovMat(sumCl      = sumCl,
+                                 timepoints = tp,
                                  sigma      = sigma,
                                  tau        = tau,
                                  eta        = eta,
@@ -559,23 +551,93 @@ compute_wlsPower <- function(DesMat,
                                  INDIV_LVL  = INDIV_LVL)
 
   ## matrices for power calculation #####
-  tmpmat <- t(dsnmatrix) %*% Matrix::chol2inv(Matrix::chol(CovMat))
-  VarMat <- Matrix::solve(tmpmat %*% dsnmatrix)
-  if(verbose>0) ProjMat <- matrix((VarMat %*% tmpmat)[1,],
-                                nrow=ifelse(INDIV_LVL,SumSubCl,SumCl),
-                                byrow=TRUE)
+  dsncols <- dim(dsn)[2]
+
+  XW  <- matrix( ((tdsn <- t(dsn)) %*% chol2inv(drop0(chol(CovMat))))@x, dsncols ) ## drop0 for incomplete designs
+  Var <- spdinv( VarInv <- XW %*% dsn )
+  if(verbose>0) ProjMat <- matrix((Var[1,] %*% XW),
+                                  nrow=ifelse(INDIV_LVL,SumSubCl,sumCl),
+                                  byrow=TRUE)
+
+  ## Information content, if requested ####
+  # currently computed twice, once explicitly, once using the specific formula
+  # explicit computation will be removed in near future
+  if(INFO_CONTENT){
+    I <- 1:sumCl
+    J <- 1:tp
+    InfoContent <- list(Cells   = matrix(NA,sumCl,tp),
+                        Cluster = numeric(sumCl),
+                        time    = numeric(tp))
+    tp_drop <- array(0,dim=c(dsncols,dsncols,tp))
+
+    if(length(CovMat@x)<tp*tp*sumCl) {  ## ugly. Is needed to produce consistent sparse matrix indexing for tau=0 (and eta >=0)
+      i <- rep(J,tp)      + (i_add <- rep(tp*(I-1),each=tp*tp) ) ## wouldnt be necessary with bdiag_m ...
+      j <- rep(J,each=tp) +  i_add
+      CovMat <- CovMat + sparseMatrix(i,j,x=0)
+    }
+
+    for(i in I){
+      J_start  <- tp*(i-1)
+      J_incomp <- if(is.null(DesMat$incompMat)) J else J[DesMat$incompMat[i,]==1]  ## no computation of empty cells
+
+      Var_drop <-spdinv( VarInv - XW[,(J_start+J)] %*% submatrix(dsn,J_start+1,J_start+tp,1,dsncols) )
+      InfoContent$Cluster[i] <- Var_drop[1,1]/Var[1,1]
+
+      if(tp>1){
+        for(j in J_incomp){
+          J_drop <- (J_ <- J[-j]) + J_start
+          XW_drop <- matrix(0, dsncols, tp)  ## add *real* zeros to remain consistent with sparse matrix indexing
+          XW_drop[,J_] <- tdsn[,J_drop] %*%
+            spdinv( matrix(CovMat@x[tp^2*(i-1) + 1:(tp)^2],tp) [J_,J_] )
+          Var_drop <- spdinv( VarInv +
+              (tp_drop_updt <- (XW_drop[,J]-XW[,J_start+J]) %*% dsn[J_start+J,]) )
+          tp_drop[,,j]           <- tp_drop[,,j] + tp_drop_updt
+          InfoContent$Cells[i,j] <- Var_drop[1,1]/Var[1,1]
+        }
+      }
+    }
+    if( tp>1 & sum(colSums(DesMat$trtMat)>0)>1 ){
+      if(DesMat$timeAdjust=="factor"){ ## TODO: ADD WARNINGS !!
+        for(j in J){
+          Var_drop <- spdinv( (VarInv + tp_drop[,,j])[-(j+1),-(j+1)] )
+          InfoContent$time[j] <- Var_drop[1,1]/Var[1,1]
+        }
+      }else {
+        for(j in J){
+          Var_drop <- spdinv( VarInv + tp_drop[,,j] )
+          InfoContent$time[j] <- Var_drop[1,1]/Var[1,1]
+        }
+      }
+    }
+
+  ## Formula-based calculation of information content
+    W  <- spdinv(as.matrix(CovMat))
+    X2 <- dsn[,-1]
+    x1 <- dsn[, 1]
+    Q  <- W %*% X2 %*% spdinv(t(X2)%*%W%*%X2) %*%t(X2)%*%W
+    WQ <- W - Q
+    hh <- as.numeric( t(x1)%*%WQ / c(t(x1)%*%WQ%*%x1) )
+    InfoContent$Closed <- matrix(1/(1 - hh^2*as.numeric(t(x1)%*%WQ%*%x1) / diag(WQ) ),
+                                 sumCl,tp, byrow=TRUE)
+
+    # ## Check consistency of methods
+    # maxDiff <- 0
+    # maxDiff <- max(abs(InfoContent$Cells - InfoContent$Closed))
+    # if(maxDiff>1e-12)  warning("formula-based information content and explicit information content differ")
+  }
+
 
   ## ddf for power calculation #####
   df <- switch(dfAdjust,
                "none"           = Inf,
-               "between-within" = SumCl - rankMatrix(dsnmatrix),
-               "containment"    = dim(dsnmatrix)[1] - SumCl,
-               "residual"       = dim(dsnmatrix)[1] - rankMatrix(dsnmatrix))
+               "between-within" = sumCl - rankMatrix(dsn),
+               "containment"    = dim(dsn)[1] - sumCl,
+               "residual"       = dim(dsn)[1] - rankMatrix(dsn))
   if(df<3){
     warning(dfAdjust,"-method not applicable. No DDF adjustment used.")
     df <- Inf }
 
-  Pwr <- tTestPwr(d=EffSize, se=sqrt(VarMat[1,1]), df=df, sig.level=sig.level)
+  Pwr <- tTestPwr(d=EffSize, se=sqrt(Var[1,1]), df=df, sig.level=sig.level)
   if(verbose==0){
     out <- Pwr
   } else {
@@ -597,12 +659,17 @@ compute_wlsPower <- function(DesMat,
                              dfAdjust   = dfAdjust,
                              sig.level  = sig.level),
                 ProjMatrix = ProjMat)
+  if(INFO_CONTENT){
+    out <- append(out,
+                  list(InformationContent= InfoContent))
   }
   if(verbose==2)
     out <- append(out,
                   list(DesignMatrix     = DesMat,
-                       CovarianceMatrix = CovMat))
+                       CovarianceMatrix = CovMat,
+                       VarianceMatrix   = Var))
   return(out)
+  }
 }
 
 #' @title Print an object of class `wlsPower`
@@ -632,17 +699,165 @@ print.wlsPower <- function(x, ...){
 
 
 
+#' @title plot the information content of a wls object
+#'
+#' @inheritParams plot.wlsPower
+#' @param IC a matrix with information content for each cluster at each time period
+#'
+#' @return a plotly object
+#' @export
+#'
+
+plot_InfoContent <- function(IC,
+                             annotations=NULL,
+                             show_colorbar=TRUE,
+                             marginal_plots=TRUE){
+
+  if(is.null(annotations)){
+    annotations <- ifelse(length(IC$Cells)<=1e2,TRUE,FALSE)
+  }
+  mx  <- max(IC$Cells)
+  sumCl <- dim(IC$Cells)[1]
+  timep <- dim(IC$Cells)[2]
+  gaps  <- 20/sqrt(length(IC$Cells))
+
+  dat=cbind(expand.grid(y=seq_len(sumCl),
+                        x=seq_len(timep)),
+            z=as.numeric(IC$Cells),
+            zChar=as.character(round(IC$Cells,3)))
+  dat$zChar[is.na(dat$zChar)] <- ""
+
+  PLT <- plot_ly(data=dat, x=~x, y=~y,  z=~z,
+          type="heatmap",
+          colors=grDevices::colorRamp(c("white","gold","firebrick")),
+          xgap=gaps, ygap=gaps, name=" ",
+          showscale=show_colorbar,
+          hovertemplate="Time: %{x}\nCluster: %{y}\nInfoContent: %{z:.6f}" ) %>%
+    colorbar(len=1,limits=c(1-1e-8,mx)) %>%
+    layout(xaxis=list(title="Time"),
+           yaxis=list(title="Cluster", autorange="reversed"))
+  if(annotations) PLT <- PLT %>% add_annotations(text=~zChar, showarrow=FALSE)
+
+
+  if(marginal_plots){
+    PLT <- subplot(
+      plot_ly(data=data.frame(time   = seq_along(IC$time),
+                              InfoC  = IC$time),
+              type="bar", x=~time, y=~(InfoC-1), color=I("grey"),
+              name=" ", base=1,
+              hovertemplate="Time: %{x}\nInfoContent: %{y:.6f}") %>%
+        layout(yaxis=list(title=""),
+               xaxis=list(title="", showticklabels=FALSE))
+      ,
+      plotly_empty(type="scatter",mode="marker")
+      ,
+      PLT
+      ,
+      plot_ly(data=data.frame(cluster = seq_along(IC$Cluster),
+                              InfoC   = IC$Cluster),
+              type="bar", orientation="h",
+              y=~cluster, x=~(InfoC-1), color=I("grey"),
+              name=" ", base=1,
+              hovertemplate="Cluster: %{y}\nInfoContent: %{x:.6f}") %>%
+        layout(xaxis=list(title=""),
+               yaxis=list(title="", showticklabels=FALSE, autorange="reversed"))
+      ,
+      nrows=2, heights=c(.2,.8), widths=c(.8,.2), titleX=TRUE, titleY=TRUE
+    ) %>% layout(showlegend=FALSE)
+  }
+  PLT
+}
+
+
+#' @title plot cell contributions (weights) of a wls object
+#'
+#' @inheritParams plot.wlsPower
+#'
+#' @return a plotly html widget
+#' @export
+#'
+plot_CellWeights <- function(x,
+                             annotations=NULL,
+                             show_colorbar=TRUE,
+                             marginal_plots=TRUE){
+
+  if(is.null(annotations)){
+    annotations <- ifelse(length(x$ProjMatrix)<=1e2,TRUE,FALSE)
+  }
+  wgt <- x$ProjMatrix
+  mx  <- max(abs(wgt))
+  sumCl <- dim(wgt)[1]
+  timep <- dim(wgt)[2]
+  gaps  <- 20/sqrt(length(wgt))
+
+  if(!is.null(x$DesignMatrix$incompMat))
+    wgt[x$DesignMatrix$incompMat==0] <- NA
+
+  dat <- cbind(expand.grid(y=seq_len(sumCl),
+                           x=seq_len(timep)),
+               wgt=as.numeric(wgt),
+               wgtChar=as.character(round(wgt,3)) )
+  dat$wgtChar[is.na(dat$wgtChar)] <- ""
+
+  PLT <- plot_ly(data=dat, x=~x, y=~y, z=~wgt, type="heatmap",
+                 colors=grDevices::colorRamp(c("steelblue","white","firebrick")),
+                 xgap=gaps, ygap=gaps, name=" ",
+                 showscale=show_colorbar,
+                 hovertemplate="Time: %{x}\nCluster: %{y}\nWeight: %{z:.6f}") %>%
+    colorbar(len=1,limits=c(-mx,mx)) %>%
+    layout(xaxis=list(title="Time"),
+           yaxis=list(title="Cluster", autorange="reversed"))
+  if(annotations) PLT <- PLT %>% add_annotations(x=dat$x, y=dat$y,
+                                                 text=dat$wgtChar, showarrow=FALSE)
+
+  if(marginal_plots){
+    PLT <- suppressWarnings(
+      subplot(
+        plot_ly(data=data.frame(time   = seq_len(timep),
+                                weight = colSums(abs(wgt),na.rm=TRUE)),
+                type="bar", x=~time, y=~weight, color=I("grey"),
+                name=" ",
+                hovertemplate="Time: %{x}\nWeight: %{y:.6f}") %>%
+          layout(yaxis=list(title="Sum|weights|"),
+                 xaxis=list(title="", showticklabels=FALSE))
+        ,
+        plotly_empty(type="scatter",mode="marker")
+        ,
+        PLT
+        ,
+        plot_ly(data=data.frame(cluster=seq_len(sumCl),
+                                weight=rowSums(abs(wgt),na.rm=TRUE)),
+                type="bar", orientation="h",
+                y=~cluster, x=~weight, color=I("grey"),
+                name=" ",
+                hovertemplate="Cluster: %{y}\nAbsWeight: %{x:.6f}") %>%
+          layout(xaxis=list(title="Sum|weights|"),
+                 yaxis=list(title="", showticklabels=FALSE, autorange="reversed"))
+        ,
+        nrows=2, heights=c(.2,.8), widths=c(.8,.2), titleX=TRUE, titleY=TRUE
+      ) %>% layout(showlegend=FALSE)
+    )
+  }
+  PLT
+}
+
+
+
 #' @title plot an object of class `wlsPower`
 #'
-#' @description Up to three plots (selectable by `which`) that visualise:
-#' influence of each cluster for each timepoint,
-#' the treatment status for each cluster for each timepoint and
-#' the covariance matrix. By default, only the first plot is returned.
+#' @description Up to four plots (selectable by `which`) that visualise:
+#' the contribution of each cluster-period cell to the treatment effect estimator,
+#' the information content of each cluster-period cell,
+#' the treatment status for each cluster for each time point and
+#' the covariance matrix. By default, only the first two plots are returned.
 #'
 #' @param x object of class wlsPower
-#' @param which Specify a subset of the numbers 1:3 to select plots
-#' @param show_colorbars logical, should the colorbars be shown?
+#' @param which Specify a subset of the numbers `1:4` to select plots. The default is
+#' `1:2` or `1`, depending on whether `x` contains the information content.
+#' @param show_colorbar logical, should the colorbars be shown?
 #' @param ... Arguments to be passed to methods
+#' @param annotations logical, should the cell contributions be annotated in the Plot?
+#' @param marginal_plots should the influence of whole periods, clusters also be plotted?
 #'
 #' @method plot wlsPower
 #'
@@ -650,53 +865,36 @@ print.wlsPower <- function(x, ...){
 #'
 #' @export
 #'
-plot.wlsPower <- function(x, which=1, show_colorbars=NULL, ...){
+plot.wlsPower <- function(x, which=NULL, show_colorbar=NULL,
+                          annotations=NULL, marginal_plots=TRUE,...){
+  out <- list()
+  if(is.null(which))
+    which <- if("InformationContent" %in% names(x)) 1:2 else 1
 
-  WgtPlot <- if (1 %in% which){
-    wgt <- x$ProjMatrix
-    mx  <- max(abs(wgt))
-    sumCl <- dim(wgt)[1]
-    timep <- dim(wgt)[2]
+  if (1 %in% which){
+    out$WgtPlot <- plot_CellWeights(x, annotations=annotations,
+                                show_colorbar=show_colorbar,
+                                marginal_plots=marginal_plots)
+  }
 
-    suppressWarnings(
-      subplot(
-        plot_ly(data=data.frame(time   = seq_len(dim(wgt)[2]),
-                                weight = colSums(abs(wgt))),
-                type="bar", x=~time, y=~weight, color=I("grey")) %>%
-          layout(yaxis=list(title="Sum|weights|"),
-                 xaxis=list(title="", showticklabels=FALSE))
-        ,
-        plotly_empty(type="scatter",mode="marker")
-        ,
-        plot_ly(x=seq_len(timep), y=seq_len(sumCl), z=wgt, type="heatmap",
-                colors=grDevices::colorRamp(c("steelblue","white","firebrick")),
-                xgap=.3, ygap=.3, name=" ",
-                hovertemplate="Time: %{x}\nCluster: %{y}\nWeight: %{z}") %>%
-          colorbar(len=1,limits=c(-mx,mx)) %>%
-          layout(xaxis=list(title="time"),
-                 yaxis=list(title="cluster", autorange="reversed"))
-        ,
-        plot_ly(data=data.frame(cluster=seq_len(dim(wgt)[1]),
-                                weight=rowSums(abs(wgt))),
-                type="bar", orientation="h",
-                y=~cluster, x=~weight, color=I("grey")) %>%
-          layout(xaxis=list(title="Sum|weights|"),
-                 yaxis=list(title="", showticklabels=FALSE, autorange="reversed"))
-        ,
-        nrows=2, heights=c(.2,.8), widths=c(.8,.2), titleX=TRUE, titleY=TRUE
-      ) %>% layout(showlegend=FALSE)
-    )
-  } else NULL
+  if (2 %in% which){
+    if(!("InformationContent" %in% names(x)) ) stop("Please rerun wlsPower() with INFO_CONTENT=TRUE")
+    out$IMplot <- plot_InfoContent(x$InformationContent,
+                                        annotations=annotations,
+                                        show_colorbar=show_colorbar,
+                                        marginal_plots=marginal_plots)
+  }
 
-  DMplot <- if (2 %in% which){
+  if (3 %in% which){
     if(!("DesignMatrix" %in% names(x)) ) stop("Please rerun wlsPower() with verbose=2")
-    plot(x$DesignMatrix, show_colorbar=show_colorbars)
-  } else NULL
+    out$DMplot <- plot(x$DesignMatrix, show_colorbar=show_colorbar)
+  }
 
-  CMplot <- if (3 %in% which){
+  if (4 %in% which){
     if(!("CovarianceMatrix" %in% names(x)) ) stop("Please rerun wlsPower() with verbose=2")
-    plot_CovMat(x$CovarianceMatrix, show_colorbar=show_colorbars)
-  } else NULL
+    out$CMplot <- plot_CovMat(x$CovarianceMatrix, show_colorbar=show_colorbar)
+  }
 
-  return(list(WgtPlot,DMplot,CMplot))
+  return(out)
 }
+